python/samples/05-observability/pyproject.toml
[project]
name = "botas-sample-05-observability"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "botas-fastapi",
    "langchain-azure-ai>=0.1.0",
    "langchain-core>=0.3.0",
    "microsoft-opentelemetry",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-instrumentation-logging>=0.43b0",
]

[tool.uv.sources]
botas = { path = "../../packages/botas", editable = true }
botas-fastapi = { path = "../../packages/botas-fastapi", editable = true }
